Fix telegram_contacts.tg_user_id int32 overflow -> BIGINT
- TG id 5506243702 exceeds int32 max -> asyncpg DataError in webhook
- models.py: tg_user_id Integer -> BigInteger
- migration 027: ALTER COLUMN TYPE BIGINT (PostgreSQL only)
- HOTFIX-3 in lifespan: converts type on startup if still integer
- test: store + lookup of 5506243702

+++ backend/app/models.py
@@ -1,6 +1,6 @@
 from datetime import datetime, timezone
 
-from sqlalchemy import JSON, Boolean, DateTime, ForeignKey, Integer, String, Text
+from sqlalchemy import JSON, BigInteger, Boolean, DateTime, ForeignKey, Integer, String, Text
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -131,7 +131,7 @@ class TelegramContact(Base):
     __tablename__ = "telegram_contacts"
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True)
-    tg_user_id: Mapped[int] = mapped_column(Integer, unique=True, nullable=False, index=True)
+    tg_user_id: Mapped[int] = mapped_column(BigInteger, unique=True, nullable=False, index=True)
     username: Mapped[str] = mapped_column(Text, nullable=False, default="")
     first_name: Mapped[str] = mapped_column(Text, nullable=False, default="")
     user_id: Mapped[int | None] = mapped_column(ForeignKey("users.id", ondelete="SET NULL"), nullable=True, index=True)
